% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPIV_test.R
\name{weak_RPIV_test}
\alias{weak_RPIV_test}
\title{Weak-IV-Robust Residual Prediction Test for Linear Instrumental Variable Models}
\usage{
weak_RPIV_test(
  Y,
  X,
  C = NULL,
  Z,
  frac_A = NULL,
  gamma = 0.05,
  variance_estimator = "heteroskedastic",
  clustering = NULL,
  upper_clip_quantile = 0.8,
  regr_par = list(),
  fit_intercept = TRUE,
  fit_at_tsls = TRUE,
  use_C_for_prediction = TRUE
)
}
\arguments{
\item{Y}{A numeric vector. The outcome variable.}

\item{X}{A numeric matrix or vector. The endogenous explanatory variables.}

\item{C}{A numeric matrix, vector or \code{NULL}. The additional exogenous explanatory variables (optional).}

\item{Z}{A numeric matrix or vector. The instruments.}

\item{frac_A}{A numeric scalar between 0 and 1 or \code{NULL}. The fraction of the sample used for training (sample splitting). Default is \code{min(0.5, exp(1)/log(n))}, where \code{n} is the sample size.}

\item{gamma}{A non-negative scalar. If the variance estimator is less than gamma times the noise level (as estimated as by the mean of the squared residuals), gamma times the noise level is used as variance estimator.}

\item{variance_estimator}{Character string or vector. One or more of \code{"homoskedastic"}, \code{"heteroskedastic"}, \code{"cluster"}. Specifies the types of variance estimation used.}

\item{clustering}{A vector of cluster identifiers or \code{NULL}. Observations with the same value of \code{clustering} belong to the same cluster. Required if \code{variance_estimator} includes \code{"cluster"}.}

\item{upper_clip_quantile}{A scalar between 0 and 1. The estimated weight-function will be clipped at the corresponding quantile of the random forest predictions on the auxiliary sample. Use 0 to use the sign of the predictions. Default is 0.8.}

\item{regr_par}{A list of parameters passed to the random forest regression model. Supports \code{num.trees}, \code{num_mtry} (number of different mtry values to try out) or a vector \code{mtry}, a vector \code{max.depth}, \code{num_min.node.size} (number of different min.node.size values to try out) or a vector \code{min.node.size}.}

\item{fit_intercept}{Logical. Should an intercept be included in the model (added to C)? Default is \code{TRUE}.}

\item{fit_at_tsls}{Logical. If \code{TRUE}, a random forest is initially tuned and fitted at the TSLS estimator. Subsequent evaluations may reuse this tuning or the learned partition structure. Default is \code{TRUE}.}

\item{use_C_for_prediction}{Logical. If \code{TRUE}, the random forest uses both \code{Z} and \code{C} as predictors. If \code{FALSE}, only Z is used. Default is \code{TRUE}.}
}
\value{
A function \code{f(beta, type)} that computes the weak RPIV test statistic.
\itemize{
\item \code{beta} is a numeric vector of coefficients.
\item \code{type} determines how the weight function is constructed:
\describe{
\item{\code{"tune_and_fit"}}{Re-tunes and refits the random forest.}
\item{\code{"fit"}}{Uses tuning parameters obtained at TSLS to fit a new random forest.}
\item{\code{"recalculate"}}{Reuses the partition obtained from the random forest with at the TSLS-residuals and recomputes leaf means.}
}
}

The returned value is a named numeric vector with one entry per variance estimator.
The resulting test statistic asymptotically follows a standard Gaussian distribution.
P-values can be obtained from the test statistic \code{T_stat} as \code{1 - pnorm(T_stat)}.
}
\description{
Constructs a hypothesis test for the well-specification of linear instrumental variable
(IV) model at a particular value of the parameter.
More specifically, it tests the null-hypothesis
\eqn{H_0(\beta_0): \mathbb E[Y-X^T\beta_0|Z] = 0.}
It uses sample splitting and a random forest to try to predict the
residuals from the instruments \eqn{Z}. The testing procedure remains valid
under weak identification. The function returns a closure that must be evaluated at a candidate parameter value.
}
\details{
The procedure is based on orthogonalized residuals and sample splitting.
A random forest is trained on an auxiliary sample to predict structural residuals
from the instruments. The resulting weight function is then partialled out with
respect to exogenous covariates and evaluated on the main sample.

Sample splitting is performed either observation-wise or at the cluster level.
Residuals and weight functions are orthogonalized with respect to exogenous
covariates to ensure validity under weak identification.
}
\examples{
set.seed(1)
n <- 100
Z <- rnorm(n)
H <- rnorm(n)
C <- rnorm(n)
X <- Z + rnorm(n) + H
Y1 <- X - C - H + rnorm(n)
Y2 <- X - C - H + rnorm(n) + Z^2
test_statistic1 <- weak_RPIV_test(Y1, X, C, Z)
test_statistic1(1, "tune_and_fit")
test_statistic1(1, "fit")
test_statistic1(1, "recalculate")
test_statistic1(0, "tune_and_fit")
test_statistic1(0, "fit")
test_statistic1(0, "recalculate")
test_statistic2 <- weak_RPIV_test(Y2, X, C, Z)
test_statistic2(1, "tune_and_fit")
test_statistic2(1, "fit")
test_statistic2(1, "recalculate")

}
\references{
Cyrill Scheidegger, Malte Londschien and Peter Bühlmann. Machine-learning-powered specification testing in linear instrumental variable models. Preprint, \url{doi:10.48550/arXiv.2506.12771}, 2025.
}
